feat(00152): implemented movie carousel

+++ 00152_movieapp/lib/presentation/themes/theme_text.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
-import 'package:movieapp/common/constants/size_constants.dart';
-import 'package:movieapp/common/extensions/size_extensions.dart';
+// ignore: import_of_legacy_library_into_null_safe
+import 'package:movieapp/common/screenutil/screnutil.dart';
 
 class ThemeText {
   ThemeText._();
@@ -9,7 +9,7 @@ class ThemeText {
   static TextTheme get _poppingsTextTheme => GoogleFonts.poppinsTextTheme();
   static TextStyle get _whiteHeadline6 =>
       _poppingsTextTheme.headline6!.copyWith(
-        fontSize: Sizes.dimen_20.sp.toDouble(),
+        fontSize: Sizes.dimen_20.sp,
         color: Colors.white,
       );
   static getTextTheme() => TextTheme(
